Introduces a collection of WIPNOTES
WIPNOTES describe things that are in progress. Such documents are likely
to be retired at some point, or they describe the system at a particular
point in time. Beware of whether the description holds currently.

The aim is for these documents to be at least as useful to me as they are
to someone else. If anything they should be more useful to me as a refresher
behind "what was I thinking when I ..."

This commit collapses a combination of commits, since they might as well be.
Includes 5 initial WIPNOTES:

.
├── 1.parser-analysis.md - parser output summmary
├── 2.pynodes.md - Representing Python Programs using PyNodes
├── 3.Python_AST_Structure.md - Python AST Structure
├── 4.CodeStructure.md - Files forming the Code structure
├── 5.ContextNodes.md - Notes on where context should and should not be
└── README

Hopefully these documents are misleading, but they are likely to be
out of date, hence called "WIPNOTES", they're notes to support current
WIP.