Transform function calls to 'print' into function print statements under the hood.
This will need to have a check at some point that the user has not redefined
'print' in the current code context. However, at present the code that is used
to do the transformation does not support that kind of check, so there will
be some work necessary there - if we care.