Move the actual entry points for compiling/etc into an API file to serve as a nascent API
Why?

Received a query about importing pyxie/core.py into a file and whether this
was the right thing, wrong thing, good thing, bad thing to do. In particlar
the query was based on a desire to use pyxie inside another project.

Pyxie's main API has always been intended to be the one exposed by the actual
front end bin/pyxie, and of course the specific language subset it exposes.
On the flipside, the internals of the pyxie.core are still changing as the
implementation improves. As a result exposing pyxie/core.py as an external
API (which then has some measure of commitment to not change) causes problems.
As a result, it made sense to move the implementation of bin/pyxie into
pyxie/api.py - so that the front end can effectively be used programmatically.

I'm not certain that the names in that file are correct. However, by creating
a hook that can be used as follows:

    from pyxie.api import initialise_API
    from pyxie.api import set_profile
    from pyxie.api import PyxieAPI

    initialise_API(profile=default)
        # Initialises the API - must be called before the other options

    set_profile(profile_name)
        # If after using the API you decide you want to change the profile, you can,
        # using this option.

    PyxieAPI.parse(filename)
        # parses the given filename, outputs result to console

    PyxieAPI.analyse(filename)
        # parses and analyses the given filename, outputs result to console

    PyxieAPI.codegen(filename, result_filename=None)
        # parses, analyse and generate code for the given filename, outputs result
        # to console. Does not attempt compiling

    PyxieAPI.compile(filename, result_filename=None)
        # compiles the given file.
        # for a given file some/program.pyxie it is compiled to some/program
        # result_filename can be provide an alternative output name

I think is a good starting point. In particular, this only really exposes 3 names that
can't change. The methods parse, analyse, codegen and compile are all independent of each
other. You can just call compile if you want. These may be augmented by better versions
as time goes on.