Remove use of thunk middleman
The middleman is the mk* functions - these allowed the use of jsonify and
iiNodeNames to co-exist.  Since the jsonify and lists are no longer used,
these middleman functions can now disappear.

There is an open question about iiOperator, but that suggests that
iiOperator is in fact badly named and can be simplified somewhat.  Perhaps
the approach to take is to have classes of operators - infix and prefix, and
then to treat these as special cases of function calls instead.

That would the simplify the implementation.

That can wait until the ECS based iiNodes though